Refactorization - First pass
Instead of one giant file now, we have 3.

One file for the main functionality - titled 'main.'

One file for the board class - stores the data about what tiles are on the board, and also calculates the best move given the current board.

And finally, one file for all the "web_stuff". By this, we mean the code necessary to navigate to the website, to read the current state of the board from the website, and to send the appropriate action to the website.